+++ onnxruntime/core/providers/tensorrt/tensorrt_execution_provider.h
@@ -89,7 +89,7 @@ struct TensorrtFuncState {
   tensorrt_ptr::unique_pointer<nvinfer1::INetworkDefinition>* network = nullptr;
   std::vector<std::unordered_map<std::string, int>> input_info;
   std::vector<std::unordered_map<std::string, int>> output_info;
-  std::unordered_map<std::string, std::unordered_map<int, std::pair<int64_t, int64_t>>> input_shape_ranges;
+  std::unordered_map<std::string, std::unordered_map<size_t, std::pair<int64_t, int64_t>>> input_shape_ranges;
   OrtMutex* tensorrt_mu_ptr = nullptr;
   bool fp16_enable;
   bool int8_enable;
@@ -175,7 +175,7 @@ class TensorrtExecutionProvider : public IExecutionProvider {
   std::unordered_map<std::string, tensorrt_ptr::unique_pointer<nvinfer1::INetworkDefinition>> networks_;
   std::unordered_map<std::string, std::vector<std::unordered_map<std::string, int>>> input_info_;
   std::unordered_map<std::string, std::vector<std::unordered_map<std::string, int>>> output_info_;
-  std::unordered_map<std::string, std::unordered_map<std::string, std::unordered_map<int, std::pair<int64_t, int64_t>>>> input_shape_ranges_;
+  std::unordered_map<std::string, std::unordered_map<std::string, std::unordered_map<size_t, std::pair<int64_t, int64_t>>>> input_shape_ranges_;
 
   /**Get IndexedSubGraph based on node list of the subgraph*/
   std::unique_ptr<IndexedSubGraph> GetSubGraph(SubGraph_t graph_nodes_index,
